[State Sync] Move SynchronizationState out of lib file.
