+++ state-synchronizer/src/lib.rs
@@ -7,9 +7,7 @@
 //! Used for node restarts, network partitions, full node syncs
 #![recursion_limit = "1024"]
 
-use diem_types::{epoch_state::EpochState, ledger_info::LedgerInfoWithSignatures};
-use executor_types::ExecutedTrees;
-pub use state_synchronizer::{StateSynchronizer, StateSynchronizerClient};
+pub use self::state_synchronizer::{StateSynchronizer, StateSynchronizerClient};
 
 pub mod chunk_request;
 pub mod chunk_response;
@@ -20,50 +18,6 @@ mod logging;
 pub mod network;
 mod request_manager;
 mod state_synchronizer;
-
-/// The state distinguishes between the following fields:
-/// * highest_local_li is keeping the latest certified ledger info
-/// * synced_trees is keeping the latest state in the transaction accumulator and state tree.
-///
-/// While `highest_local_li` can be used for helping the others (corresponding to the highest
-/// version we have a proof for), `synced_trees` is used for retrieving missing chunks
-/// for the local storage.
-#[derive(Clone)]
-pub struct SynchronizationState {
-    pub highest_local_li: LedgerInfoWithSignatures,
-    synced_trees: ExecutedTrees,
-    // Corresponds to the current epoch if the highest local LI is in the middle of the epoch,
-    // or the next epoch if the highest local LI is the final LI in the current epoch.
-    pub trusted_epoch: EpochState,
-}
-
-impl SynchronizationState {
-    pub fn new(
-        highest_local_li: LedgerInfoWithSignatures,
-        synced_trees: ExecutedTrees,
-        current_epoch_state: EpochState,
-    ) -> Self {
-        let trusted_epoch = highest_local_li
-            .ledger_info()
-            .next_epoch_state()
-            .cloned()
-            .unwrap_or(current_epoch_state);
-        SynchronizationState {
-            highest_local_li,
-            synced_trees,
-            trusted_epoch,
-        }
-    }
-
-    /// The highest available version in the local storage (even if it's not covered by the LI).
-    pub fn highest_version_in_local_storage(&self) -> u64 {
-        self.synced_trees.version().unwrap_or(0)
-    }
-
-    pub fn epoch(&self) -> u64 {
-        self.trusted_epoch.epoch
-    }
-}
 
 #[cfg(any(feature = "fuzzing", test))]
 mod tests;
