+++ state-synchronizer/src/tests/mock_storage.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::SynchronizationState;
+use crate::state_synchronizer::SynchronizationState;
 use anyhow::{bail, Result};
 use diem_crypto::HashValue;
 #[cfg(test)]
